batch: Move content commit helpers
Batch storage delegates realized text construction to a named module. It
also contains one-path commit tree publication helpers for binary, gitlink,
and sift flows.

Those helpers are raw Git tree mutation operations rather than file
metadata persistence. Keeping them on storage makes command code depend on
the broad storage module when it only needs content commit publication.

This commit continues separating batch persistence by moving the commit
parent calculation and one-path tree mutations into batch.content_commits,
then updating production callers.

Subsequent commits will move architecture checks to the new owner.
